+++ modules/client/src/tokenVoting/index.ts
@@ -7,4 +7,6 @@ export {
   TokenVotingProposal,
   TokenVotingProposalListItem,
   TokenVotingProposalResult,
+  WrapTokensStep,
+  UnwrapTokensStep
 } from "./interfaces";